rename commands namespace (bc break)
