+++ src/Commands/ICommandHandler.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace Damejidlo\CommandBus;
+namespace Damejidlo\Commands;
 
 use Damejidlo\MessageBus\IMessageHandler;
