+++ tests/Integration/CommandHandlingTest.phpt
@@ -9,7 +9,7 @@ namespace DamejidloTests\Integration;
 
 require_once __DIR__ . '/../bootstrap.php';
 
-use Damejidlo\CommandBus\NewEntityId;
+use Damejidlo\Commands\NewEntityId;
 use Damejidlo\Events\IEventDispatcher;
 use Damejidlo\MessageBus\Handling\HandlerCannotBeProvidedException;
 use Damejidlo\MessageBus\Handling\HandlerRequiredAndNotConfiguredException;
