+++ tests/Integration/Fixtures/PlaceOrderCommand.php
@@ -2,7 +2,7 @@
 
 namespace DamejidloTests\Integration\Fixtures;
 
-use Damejidlo\CommandBus\ICommand;
+use Damejidlo\Commands\ICommand;
