+++ tests/Integration/SynchronousEventsDispatchingTest.phpt
@@ -9,9 +9,9 @@ namespace DamejidloTests\Integration;
 
 require_once __DIR__ . '/../bootstrap.php';
 
-use Damejidlo\CommandBus\CommandBus;
-use Damejidlo\CommandBus\ICommandBus;
-use Damejidlo\CommandBus\NewEntityId;
+use Damejidlo\Commands\CommandBus;
+use Damejidlo\Commands\ICommandBus;
+use Damejidlo\Commands\NewEntityId;
 use Damejidlo\Events\CommandBusAwareEventDispatcher;
 use Damejidlo\Events\EventBus;
 use Damejidlo\Events\InMemoryEventQueue;
