+++ tests/MessageBus/Logging/LogMessageResolverTest.phpt
@@ -5,7 +5,7 @@ namespace DamejidloTests\MessageBus\Logging;
 
 require_once __DIR__ . '/../../bootstrap.php';
 
-use Damejidlo\CommandBus\ICommand;
+use Damejidlo\Commands\ICommand;
 use Damejidlo\Events\IDomainEvent;
 use Damejidlo\MessageBus\Handling\HandlerType;
 use Damejidlo\MessageBus\IMessage;
